Add domain for setup_fs.
This oneshot service was previously left in init's domain.
It requires access to initialize userdata and cache, so label the specific
block devices with specific types so that we can allow access without
allowing access to any other block devices.

Depends on a change to external/sepolicy with the same Change-Id to
define the userdata and cache block device types.

Change-Id: I99d24f06506f51ebf1d186d9c393b3cad60e98d7
Signed-off-by: Stephen Smalley <sds@tycho.nsa.gov>